move handlers module to separate file in src/server/api
